+++ examples/robot/ragtime_panthera/follower/main/include/app_console.h
@@ -12,8 +12,9 @@
  * @brief Control panthera using console commands
  *
  * @param motor_control Motor control instance
+ * @param app_manager App manager instance
  * @return esp_err_t ESP_OK if successful, otherwise ESP_ERR_INVALID_ARG
  */
-esp_err_t app_console_init(damiao::Motor_Control* motor_control);
+esp_err_t app_console_init(damiao::Motor_Control* motor_control, Manager* app_manager);
 
 #endif