+++ examples/robot/ragtime_panthera/follower/CHANGELOG.md
@@ -1,5 +1,17 @@
 # ChangeLog
 
+## v0.2.0 - 2026-2-1
+
+### Enhancements:
+
+* Support switching color detection HSV range via console command
+
+### Bug Fix:
+
+* Fix `grasp_task` hanging when executing grasp task with motors not powered on
+* Fix issue where `grasp_task` state is not reset to IDLE when motors are disabled
+* Prevent SYNC switch from being toggled when motors are not enabled
+
 ## v0.1.0 - 2025-12-3
 
 ### Enhancements: